requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.2.0
django-heroku==0.3.1
django-tinymce4-lite==1.7.5
django-webpush==0.3.1
gcloud==0.17.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
http-ece==1.1.0
httplib2==0.14.0
idna==2.8
jsmin==2.2.2
jws==0.1.3
oauth2client==3.0.0
Pillow==6.2.1
protobuf==3.11.1
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2019.3
pywebpush==1.9.4
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
